132097675: FOLDER PLUGIN: Items are not clickable and, therefore, not navigable on the CP
