+++ control/design/index.html
@@ -3,17 +3,15 @@
 <head lang="en">
     <title>design</title>
     <meta charset="UTF-8">
-
     <!-- CSS -->
     <link href="../../../../styles/helper.css" rel="stylesheet">
     <link href="../../../../styles/siteIcons.css" rel="stylesheet">
-
     <!-- JS -->
     <script src="../../../../scripts/buildfire.js"></script>
     <script src="../../../../scripts/angular/angular.min.js"></script>
-    <script src="../shared.js"></script>
+    <script src="../app.services.js"></script>
+    <script src="app.js"></script>
     <script src="design.controller.js"></script>
-
 </head>
 <body ng-controller="folderPluginCtrl" ng-cloak ng-show="datastoreInitialized">
     <div ng-form="frmMain">
@@ -37,7 +35,7 @@
                 </div>
             </div>
         </div>
-		<hr>
+        <hr>
         <div class="item clearfix row margin-bottom-twenty">
             <div class="labels col-md-3 padding-right-zero pull-left">
                 <span>Background Image</span>
@@ -55,27 +53,26 @@
                 </div>
             </div>
         </div>
-
-		<div>
-			<hr>
-			<div class="item row clearfix">
-			    <div class="labels col-md-3 padding-right-zero pull-left">
-			        <span>Hide Text</span>
-			    </div>
-			    <div class="main col-md-9 pull-right" ng-init="data.design.hideText">
-					<div class="radio radio-primary radio-inline">
-						<input ng-if="data.design.hideText" id="showText" type="radio" name="toggleText"  ng-model="data.design.hideText" ng-value="false">
-                        <input ng-if="!data.design.hideText" id="showText" type="radio" name="toggleText"   checked="checked" >
+        <div>
+            <hr>
+            <div class="item row clearfix">
+                <div class="labels col-md-3 padding-right-zero pull-left">
+                    <span>Hide Text</span>
+                </div>
+                <div class="main col-md-9 pull-right" ng-init="data.design.hideText">
+                    <div class="radio radio-primary radio-inline">
+                        <input ng-if="data.design.hideText" id="showText" type="radio" name="toggleText" ng-model="data.design.hideText" ng-value="false">
+                        <input ng-if="!data.design.hideText" id="showText" type="radio" name="toggleText" checked="checked">
                         <label for="showText">False</label>
-					</div>
-					<div class="radio radio-primary radio-inline">
-						<input id="hideText" type="radio" name="toggleText" ng-model="data.design.hideText" ng-value="true">
-						<label for="hideText">True</label>
-					</div>
-					<p class="info margin-top-fifteen">Note: This will allow you to hide the text on specific layouts.</p>
-				</div>
-			</div>
-		</div>
+                    </div>
+                    <div class="radio radio-primary radio-inline">
+                        <input id="hideText" type="radio" name="toggleText" ng-model="data.design.hideText" ng-value="true">
+                        <label for="hideText">True</label>
+                    </div>
+                    <p class="info margin-top-fifteen">Note: This will allow you to hide the text on specific layouts.</p>
+                </div>
+            </div>
+        </div>
     </div>
 </body>
 </html>
