+++ widget/app.js
@@ -1,0 +1,6 @@
+(function (angular) {
+    "use strict";
+    //created folderPluginWidget module
+    angular.module('folderPluginWidget', ['infinite-scroll'])
+})
+(window.angular);
